ai/pyproject.toml
[project]
name = "vmaf-train"
version = "0.1.0"
description = "Train, export and register tiny perceptual-quality models for libvmaf."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "BSD-3-Clause-Plus-Patent" }
authors = [
  { name = "Lusoris", email = "lusoris@pm.me" },
  { name = "Claude (Anthropic)", email = "noreply@anthropic.com" },
]
dependencies = [
  "torch>=2.4,<2.6",
  "lightning>=2.4,<2.5",
  "onnx>=1.17,<2.0",
  "onnxruntime>=1.20,<2.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.2,<3.0",
  "pyarrow>=17.0",
  "scipy>=1.13",
  "scikit-learn>=1.5",
  "pyyaml>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "tqdm>=4.66",
]

[project.optional-dependencies]
tune = ["optuna>=4.0", "ray[tune]>=2.35"]
gpu  = []
viz  = ["matplotlib>=3.9", "seaborn>=0.13"]
dev  = ["pytest>=8", "ruff>=0.6", "mypy>=1.10"]

[project.scripts]
vmaf-train = "vmaf_train.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vmaf_train"]

[tool.hatch.build.targets.wheel.force-include]
"src/vmaf_train/data/manifests" = "vmaf_train/data/manifests"
"src/vmaf_train/configs" = "vmaf_train/configs"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "SIM", "RUF"]
